requirements.txt
asgiref==3.3.1
Django==3.1.6
django-appconf==1.0.4
django-debug-toolbar==3.2
django-imagefit==0.6.2
django-solo==1.1.5
django-tinymce==3.2.0
djangorestframework==3.12.2
Pillow==8.1.0
pkg-resources==0.0.0
pytz==2020.4
sqlparse==0.4.1
